requirements.txt
acdh-cidoc-pyutils
acdh-tei-pyutils
python-slugify